+++ src/lowrank/lrthresh.h
@@ -22,16 +22,16 @@ extern const struct operator_p_s* lrthresh_create(const long dims_lev[DIMS], _Bo
 extern float lrnucnorm(const struct operator_p_s* op, const complex float* src);
 
 // Generates multiscale block sizes
-extern long multilr_blkdims(long blkdims[MAX_LEV][DIMS], unsigned long flags, const long dims[DIMS], int blkskip, long initblk);
+extern int multilr_blkdims(long blkdims[MAX_LEV][DIMS], unsigned long flags, const long dims[DIMS], int blkskip, int initblk);
 
 // Generates locally low rank block size
-extern long llr_blkdims(long blkdims[MAX_LEV][DIMS], unsigned long flags, const long dims[DIMS], long llrblk);
+extern int llr_blkdims(long blkdims[MAX_LEV][DIMS], unsigned long flags, const long dims[DIMS], int llrblk);
 
 // Generates low rank plus sparse block size
-extern long ls_blkdims(long blkdims[MAX_LEV][DIMS], const long dims[DIMS]);
+extern int ls_blkdims(long blkdims[MAX_LEV][DIMS], const long dims[DIMS]);
 
 
-extern void add_lrnoiseblk(long* level, long blkdims[MAX_LEV][DIMS], const long dims[DIMS]);
+extern void add_lrnoiseblk(int* level, long blkdims[MAX_LEV][DIMS], const long dims[DIMS]);
 
 // Return the regularization parameter
 extern float get_lrthresh_lambda(const struct operator_p_s* o);
